components: split components into individual files
React conventions typically recommend having a single component
per-file, to make testing of those files (and understanding their
relationships) simpler. Split page-like components into the src/pages/
tree and other components into the src/components/ tree.
